Adding comments Adding the certificate observability as experimental Moved the CertificateRotationTests to TokenAcquirerTests (from Ms.Id.Web.Test.Integration) and use Microsoft.Identity.Web.GraphServiceClient instead of Microsoft.Identity.Web.MicrosoftGraph (because we need Graph 5 for the CertificateRotationTests)
