+++ tests/Microsoft.Identity.Web.Test.Integration/Microsoft.Identity.Web.Test.Integration.csproj
@@ -18,10 +18,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\IntegrationTests\TokenAcquirerTests\IgnoreOnAzureDevOpsFactAttribute.cs" Link="IgnoreOnAzureDevOpsFactAttribute.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="$(MicrosoftAspNetCoreMvcTestingVersion)" />
     <PackageReference Include="Microsoft.Identity.Lab.Api" Version="$(MicrosoftIdentityLabApiVersion)" />
     <PackageReference Include="NSubstitute" Version="$(NSubstituteVersion)" />
@@ -32,8 +28,6 @@
     </PackageReference>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="$(MicrosoftNetTestSdkVersion)" />
     <PackageReference Include="Newtonsoft.Json" Version="$(NewtonsoftJsonVersion)" />
-    <PackageReference Include="Microsoft.Graph" Version="5.12.0" />
-
   </ItemGroup>
 
   <ItemGroup>
